+++ src/components/Box/RvBox.tsx
@@ -4,6 +4,8 @@ import type { FC } from "react";
 import type { RvBoxFCType } from "../../types/components/Div";
 import { StyledDiv } from "./style";
 
+const rvBoxProps: RvBoxFCType = {};
+
 const RvBox: FC<RvBoxFCType> = ({ children, ...props }) => {
 	const newProps = { ...props };
 	return <StyledDiv {...newProps}>{children}</StyledDiv>;