+++ src/components/Button/style/icon.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+import StyledButton from "./base";
+
+const StyledIconButton = styled(StyledButton)`
+	display: inline-flex;
+	align-items: center;
+	justify-content: center;
+`;
+
+export default StyledIconButton;